Refactor: EditRecord -> EditObservation
