+++ gnd/src/main/res/layout/multiple_choice_input_field.xml
@@ -22,16 +22,16 @@
   <data>
     <variable
         name="fragment"
-        type="com.google.android.gnd.ui.editrecord.EditRecordFragment"/>
+        type="com.google.android.gnd.ui.editobservation.EditObservationFragment"/>
     <variable
         name="field"
         type="com.google.android.gnd.model.form.Field"/>
     <variable
         name="viewModel"
-        type="com.google.android.gnd.ui.editrecord.EditRecordViewModel"/>
+        type="com.google.android.gnd.ui.editobservation.EditObservationViewModel"/>
   </data>
 
-  <com.google.android.gnd.ui.editrecord.MultipleChoiceFieldLayout
+  <com.google.android.gnd.ui.editobservation.MultipleChoiceFieldLayout
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
       android:layout_marginTop="@dimen/field_vertical_margin"
@@ -86,5 +86,5 @@
         android:alpha="0"
         android:clickable="true"/>
 
-  </com.google.android.gnd.ui.editrecord.MultipleChoiceFieldLayout>
+  </com.google.android.gnd.ui.editobservation.MultipleChoiceFieldLayout>
 </layout>
