+++ gnd/src/main/res/layout/text_input_field.xml
@@ -26,7 +26,7 @@
         type="com.google.android.gnd.model.form.Field"/>
     <variable
         name="viewModel"
-        type="com.google.android.gnd.ui.editrecord.EditRecordViewModel"/>
+        type="com.google.android.gnd.ui.editobservation.EditObservationViewModel"/>
   </data>
 
   <LinearLayout
